refactor: split models.rs into submodules by domain
- Moved WASM-related models into models/wasm.rs
- Converted models.rs into module declaration with re-exports
- Fixed clippy warnings in conductor.rs (useless conversions, match to if let)
- Conductor models remain in conductor.rs where they logically belong
